+++ src/stories/Input.stories.tsx
@@ -44,64 +44,55 @@ InputCurrency.args = {
   endAdornment: '$',
   startAdornment: '%',
   className: 'w-full',
-  variant: 'active',
   type: 'currency'
 }
 export const InputPassword = Template.bind({})
 InputPassword.args = {
   label: 'Password',
   type: 'password',
-  className: 'w-full',
-  variant: 'active'
+  className: 'w-full'
 }
 export const InputEmail = Template.bind({})
 InputEmail.args = {
   label: 'Email',
   type: 'email',
-  className: 'w-full',
-  variant: 'active'
+  className: 'w-full'
 }
 export const InputDate = Template.bind({})
 InputDate.args = {
   label: 'Date',
   type: 'date',
-  className: 'w-full',
-  variant: 'active'
+  className: 'w-full'
 }
 export const InputMonth = Template.bind({})
 InputMonth.args = {
   label: 'Day',
   type: 'month',
-  className: 'w-full',
-  variant: 'active'
+  className: 'w-full'
 }
 export const InputYear = Template.bind({})
 InputYear.args = {
   label: 'Year',
   type: 'year',
-  className: 'w-full',
-  variant: 'active'
+  className: 'w-full'
 }
 export const InputPercentage = Template.bind({})
 InputPercentage.args = {
   label: 'Percentage',
   type: 'percentage',
-  className: 'w-56',
-  variant: 'active'
+  className: 'w-56'
 }
 
 export const InputNumber = Template.bind({})
 InputNumber.args = {
   label: 'Number',
   type: 'number',
-  className: 'w-56',
-  variant: 'active'
+  className: 'w-56'
 }
 
 export const InputFile = Template.bind({})
 InputFile.args = {
   label: 'Upload your documentation',
   type: 'file',
-  className: 'w-80',
-  variant: 'active'
+  className: 'w-80'
 }